Add files via upload
Cambios en Lancha, Helicóptero, OPVLigero, OPV

+++ ude/proyecto3/Servidor/Logica/Helicoptero.java
@@ -4,7 +4,7 @@ import ude.proyecto3.Servidor.Logica.Vehiculo;
 
 public class Helicoptero extends Vehiculo {
 	
-	public Helicoptero(String id, long angulo, long rotacion, long posx, long posy)
+	public Helicoptero(String id, float angulo, float rotacion, float posx, float posy)
 	{
 		super (id, angulo, rotacion, posx, posy);		
 	}	// Helicoptero
