+++ ude/proyecto3/Servidor/Logica/Lancha.java
@@ -4,7 +4,7 @@ import ude.proyecto3.Servidor.Logica.Vehiculo;
 
 public class Lancha extends Vehiculo {
 	
-	public Lancha(String id, long angulo, long rotacion, long posx, long posy)
+	public Lancha(String id, float angulo, float rotacion, float posx, float posy)
 	{
 		super (id, angulo, rotacion, posx, posy);		
 	}	// Lancha
