+++ ude/proyecto3/Servidor/Logica/OPVLigero.java
@@ -2,7 +2,7 @@ package ude.proyecto3.Servidor.Logica;
 
 public class OPVLigero extends OPV {
 	
-	public OPVLigero(String id, long angulo, long rotacion, long posx, long posy, int ene)
+	public OPVLigero(String id, float angulo, float rotacion, float posx, float posy, int ene)
 	{
 		super (id, angulo, rotacion, posx, posy, ene);
 	}	// OPVligero
